requirements.txt
# Core
django==4.2.8
python-dotenv==1.0.0
djangorestframework==3.15.2
djoser==2.2.3
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2

# Deployment
whitenoise==6.5.0
gunicorn==21.2.0
django-debug-toolbar==4.2.0
